+++ apps/explorer/lib/explorer/chain/address.ex
@@ -25,7 +25,7 @@ defmodule Explorer.Chain.Address do
 
   alias Explorer.Chain.Cache.NetVersion
 
-  @optional_attrs ~w(contract_code fetched_coin_balance fetched_coin_balance_block_number nonce decompiled verified)a
+  @optional_attrs ~w(contract_code fetched_coin_balance fetched_coin_balance_block_number nonce decompiled verified gas_used)a
   @required_attrs ~w(hash)a
   @allowed_attrs @optional_attrs ++ @required_attrs
 
@@ -58,7 +58,8 @@ defmodule Explorer.Chain.Address do
           contracts_creation_transaction: %Ecto.Association.NotLoaded{} | Transaction.t(),
           inserted_at: DateTime.t(),
           updated_at: DateTime.t(),
-          nonce: non_neg_integer() | nil
+          nonce: non_neg_integer() | nil,
+          gas_used: non_neg_integer() | nil
         }
 
   @derive {Poison.Encoder,
@@ -95,6 +96,7 @@ defmodule Explorer.Chain.Address do
     field(:verified, :boolean, default: false)
     field(:has_decompiled_code?, :boolean, virtual: true)
     field(:stale?, :boolean, virtual: true)
+    field(:gas_used, :integer)
 
     has_one(:smart_contract, SmartContract)
     has_one(:token, Token, foreign_key: :contract_address_hash)